refactor: Remove missing import handling for state backends now that they are loaded dynamically
Take 2 for https://github.com/meltano/meltano/pull/9475.
