+++ src/meltano/core/state_store/azure/__init__.py
@@ -1,0 +1,1 @@
+"""Azure state store."""
