+++ src/meltano/core/state_store/google/__init__.py
@@ -1,0 +1,1 @@
+"""Google state store."""
