+++ src/meltano/core/state_store/s3/__init__.py
@@ -1,0 +1,1 @@
+"""S3 state store."""
