+++ tests/meltano/core/state_store/test_filesystem.py
@@ -27,13 +27,13 @@ from botocore.stub import Stubber
 from google.cloud.storage import Blob, Bucket
 
 from meltano.core.state_store import MeltanoState
-from meltano.core.state_store.azure import AZStorageStateStoreManager
+from meltano.core.state_store.azure.backend import AZStorageStateStoreManager
 from meltano.core.state_store.filesystem import (
     _LocalFilesystemStateStoreManager,
     _WindowsFilesystemStateStoreManager,
 )
-from meltano.core.state_store.google import GCSStateStoreManager
-from meltano.core.state_store.s3 import S3StateStoreManager
+from meltano.core.state_store.google.backend import GCSStateStoreManager
+from meltano.core.state_store.s3.backend import S3StateStoreManager
 
 if t.TYPE_CHECKING:
     from collections.abc import Iterator
@@ -334,7 +334,7 @@ class TestAZStorageStateStoreManager:
     @pytest.fixture
     def mock_client(self):
         with patch(
-            "meltano.core.state_store.azure.BlobServiceClient",
+            "meltano.core.state_store.azure.backend.BlobServiceClient",
         ) as mock_client:
             yield mock_client
 
@@ -413,7 +413,7 @@ class TestS3StateStoreManager:
     @contextmanager
     def stubber(self) -> Iterator[Stubber]:
         with patch(
-            "meltano.core.state_store.s3.S3StateStoreManager.client",
+            "meltano.core.state_store.s3.backend.S3StateStoreManager.client",
             new_callable=PropertyMock,
         ) as mock_client:
             mock_client.return_value = client("s3")
